Refactor all rules into single package
