+++ src/main/java/edu/usf/cutr/gtfsrtvalidator/background/BackgroundTask.java
@@ -24,11 +24,8 @@ import edu.usf.cutr.gtfsrtvalidator.api.resource.GtfsFeed;
 import edu.usf.cutr.gtfsrtvalidator.db.GTFSDB;
 import edu.usf.cutr.gtfsrtvalidator.helper.DBHelper;
 import edu.usf.cutr.gtfsrtvalidator.helper.ErrorListHelperModel;
-import edu.usf.cutr.gtfsrtvalidator.validation.entity.StopTimeSequenceValidator;
-import edu.usf.cutr.gtfsrtvalidator.validation.entity.StopValidator;
-import edu.usf.cutr.gtfsrtvalidator.validation.entity.VehicleValidator;
-import edu.usf.cutr.gtfsrtvalidator.validation.entity.combined.*;
 import edu.usf.cutr.gtfsrtvalidator.validation.interfaces.FeedEntityValidator;
+import edu.usf.cutr.gtfsrtvalidator.validation.rules.*;
 import org.apache.commons.io.IOUtils;
 import org.hibernate.Session;
 import org.onebusaway.gtfs.impl.GtfsDaoImpl;
