+++ src/main/java/edu/usf/cutr/gtfsrtvalidator/validation/rules/FrequencyTypeZeroValidator.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.usf.cutr.gtfsrtvalidator.validation.entity.combined;
+package edu.usf.cutr.gtfsrtvalidator.validation.rules;
 
 import com.google.transit.realtime.GtfsRealtime;
 import edu.usf.cutr.gtfsrtvalidator.api.model.MessageLogModel;
