+++ src/test/java/edu/usf/cutr/gtfsrtvalidator/test/rules/FrequencyTypeZeroValidatorTest.java
@@ -3,7 +3,7 @@ package edu.usf.cutr.gtfsrtvalidator.test.rules;
 import com.google.transit.realtime.GtfsRealtime;
 import edu.usf.cutr.gtfsrtvalidator.test.FeedMessageTest;
 import edu.usf.cutr.gtfsrtvalidator.test.util.TestUtils;
-import edu.usf.cutr.gtfsrtvalidator.validation.entity.combined.FrequencyTypeZeroValidator;
+import edu.usf.cutr.gtfsrtvalidator.validation.rules.FrequencyTypeZeroValidator;
 import org.junit.Test;
 
 import java.io.IOException;
