+++ src/test/java/edu/usf/cutr/gtfsrtvalidator/test/rules/StopTimeSequenceValidatorTest.java
@@ -20,7 +20,7 @@ import com.google.transit.realtime.GtfsRealtime;
 import edu.usf.cutr.gtfsrtvalidator.test.FeedMessageTest;
 import edu.usf.cutr.gtfsrtvalidator.test.util.TestUtils;
 import edu.usf.cutr.gtfsrtvalidator.validation.ValidationRules;
-import edu.usf.cutr.gtfsrtvalidator.validation.entity.StopTimeSequenceValidator;
+import edu.usf.cutr.gtfsrtvalidator.validation.rules.StopTimeSequenceValidator;
 import org.junit.Test;
 
 import static edu.usf.cutr.gtfsrtvalidator.util.TimestampUtils.MIN_POSIX_TIME;
