+++ src/test/java/edu/usf/cutr/gtfsrtvalidator/test/rules/TripDescriptorValidatorTest.java
@@ -20,7 +20,7 @@ import com.google.transit.realtime.GtfsRealtime;
 import edu.usf.cutr.gtfsrtvalidator.test.FeedMessageTest;
 import edu.usf.cutr.gtfsrtvalidator.test.util.TestUtils;
 import edu.usf.cutr.gtfsrtvalidator.validation.ValidationRules;
-import edu.usf.cutr.gtfsrtvalidator.validation.entity.combined.TripDescriptorValidator;
+import edu.usf.cutr.gtfsrtvalidator.validation.rules.TripDescriptorValidator;
 import org.junit.Test;
 
 import static edu.usf.cutr.gtfsrtvalidator.util.TimestampUtils.MIN_POSIX_TIME;
