+++ src/test/java/edu/usf/cutr/gtfsrtvalidator/test/rules/VehicleValidatorTest.java
@@ -20,7 +20,7 @@ import com.google.transit.realtime.GtfsRealtime;
 import edu.usf.cutr.gtfsrtvalidator.test.FeedMessageTest;
 import edu.usf.cutr.gtfsrtvalidator.test.util.TestUtils;
 import edu.usf.cutr.gtfsrtvalidator.validation.ValidationRules;
-import edu.usf.cutr.gtfsrtvalidator.validation.entity.VehicleValidator;
+import edu.usf.cutr.gtfsrtvalidator.validation.rules.VehicleValidator;
 import org.junit.Test;
 import org.locationtech.spatial4j.context.jts.JtsSpatialContext;
 import org.locationtech.spatial4j.io.ShapeIO;
